+++ Timeline/ClientApp/src/app/todo/todo-item.ts
@@ -1,0 +1,6 @@
+export interface TodoItem {
+  number: number;
+  title: string;
+  isClosed: boolean;
+  detailUrl: string;
+}
