+++ Timeline/ClientApp/src/app/todo/todo-item/todo-item.component.ts
@@ -1,10 +1,10 @@
 import { Component, Input } from '@angular/core';
-import { TodoItem } from '../todo-list-page/todo-list.service';
+import { TodoItem } from '../todo-item';
 
 @Component({
   selector: 'app-todo-item',
   templateUrl: './todo-item.component.html',
-  styleUrls: ['./todo-item.component.css', '../todo-list-page/todo-list-color-block.css']
+  styleUrls: ['./todo-item.component.css', '../todo-list-color-block.css']
 })
 export class TodoItemComponent {
