+++ Timeline/ClientApp/src/app/todo/todo-list-page/todo-list-page.component.ts
@@ -1,11 +1,12 @@
 import { Component, OnInit } from '@angular/core';
-import { TodoListService, TodoItem } from './todo-list.service';
+import { TodoItem } from '../todo-item';
+import { TodoListService } from '../todo-service/todo-list.service';
 import { trigger, transition, style, animate } from '@angular/animations';
 
 @Component({
   selector: 'app-todo-list-page',
   templateUrl: './todo-list-page.component.html',
-  styleUrls: ['./todo-list-page.component.css', './todo-list-color-block.css'],
+  styleUrls: ['./todo-list-page.component.css', '../todo-list-color-block.css'],
   animations: [
     trigger('itemEnter', [
       transition(':enter', [
