+++ Timeline/ClientApp/src/app/user/user-info.ts
@@ -1,0 +1,4 @@
+export interface UserInfo {
+  username: string;
+  roles: string[];
+}
